api-resolutor/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
slowapi